+++ accessory/characteristic/BinarySensor.js
@@ -8,10 +8,10 @@ const BINARY_CONFIG = {
 };
 
 // This function will try and add a battery warning characteristic to the provided service
-function addBinarySensorCharacteristic(service, characteristic, optional) {
+function addBinarySensorCharacteristic(service, characteristic, CONFIG, optional) {
     try {
-        let [item] = this._getAndCheckItemType(BINARY_CONFIG.item, ['Contact', 'Switch']);
-        let inverted = this._checkInvertedConf(BINARY_CONFIG.inverted);
+        let [item] = this._getAndCheckItemType(CONFIG.item, ['Contact', 'Switch']);
+        let inverted = this._checkInvertedConf(CONFIG.inverted);
 
         this._log.debug(`Creating binary sensor characteristic for ${this.name} with item ${item} and inverted set to ${inverted}`);
 
@@ -43,35 +43,43 @@ function addBinarySensorCharacteristic(service, characteristic, optional) {
 }
 
 function addCarbonDioxideDetectedCharacteristic(service) {
-    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.CarbonDioxideDetected));
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.CarbonDioxideDetected), BINARY_CONFIG);
 }
 
 function addContactSensorCharacteristic(service) {
-    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.ContactSensorState));
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.ContactSensorState), BINARY_CONFIG);
 }
 
 function addCarbonMonoxideDetectedCharacteristic(service) {
-    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.CarbonMonoxideDetected));
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.CarbonMonoxideDetected), BINARY_CONFIG);
 }
 
 function addFilterChangeIndicationCharacteristic(service) {
-    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.FilterChangeIndication));
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.FilterChangeIndication), BINARY_CONFIG);
 }
 
 function addLeakDetectedCharacteristic(service) {
-    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.LeakDetected));
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.LeakDetected), BINARY_CONFIG);
 }
 
 function addMotionDetectedCharacteristic(service) {
-    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.MotionDetected));
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.MotionDetected), BINARY_CONFIG);
 }
 
 function addOccupancyDetectedCharacteristic(service) {
-    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.OccupancyDetected));
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.OccupancyDetected), BINARY_CONFIG);
 }
 
 function addSmokeDetectedCharacteristic(service) {
-    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.SmokeDetected));
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.SmokeDetected), BINARY_CONFIG);
+}
+
+function addSwingModeCharacteristic(service, optional) {
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.SwingMode), {item: "swingItem", inverted: "swingItemInverted"}, optional);
+}
+
+function addActiveCharacteristic(service, optional) {
+    addBinarySensorCharacteristic.bind(this)(service, service.getCharacteristic(this.Characteristic.Active), {item: "activeItem", inverted: "activeItemInverted"}, optional);
 }
 
 module.exports = {
@@ -82,5 +90,7 @@ module.exports = {
     addLeakDetectedCharacteristic,
     addMotionDetectedCharacteristic,
     addOccupancyDetectedCharacteristic,
-    addSmokeDetectedCharacteristic
+    addSmokeDetectedCharacteristic,
+    addSwingModeCharacteristic,
+    addActiveCharacteristic
 };